defconfig: switch to Image.gz The boot partition of the sweet device is the Header v2 version, the kernel and dtb are independent of each other, and dtb has another address of its own.
Before the header v2 version, dtb was appended to the kernel and shared
the same block address with the kernel.

PS: According to Google's requirements, for devices with Android 10 pre-installed at the factory, the boot partition must be the header
v2 version.

Therefore, for sweet devices, when making AnyKernel3 package, the kernel image (Image or Image.gz) and dtb file should be placed
separately. Only by doing this can the kernel and dtb in the boot partition be replaced.
If you still use the Image*-dtb format, then after install the kernel, the dtb in the boot partition will not be replaced, and the dtb
of appended to the kernel will not be recognized by bootloader.

You may ask: I still use the Image*-dtb format, but the user's device
can still boot after installing my kernel and there is no bugs.
Why?
Because the dtb in the kernel is provided by the Soc manufacturer (Qualcomm), and the device manufacturer (Xiaomi) changes to the dtb are
placed in the dtbo partition. So in theory, there is not much difference in the dtb source code for different third-party kernel
projects, so it is not surprising even if there is no problem.

[Credits - Pzqqt]

Signed-off-by: Palakurthi Sreekanth <palakurthisreekanth995@gmail.com>
Signed-off-by: rockstar5495 <pratyakshom.05@gmail.com>